+++ docs/API.rst
@@ -6,14 +6,12 @@ API
 
     polars_xdt.date_range
     polars_xdt.workday_count
-    polars_xdt.ExprXDTNamespace.base_utc_offset
-    polars_xdt.ExprXDTNamespace.ceil
-    polars_xdt.ExprXDTNamespace.day_name
-    polars_xdt.ExprXDTNamespace.dst_offset
-    polars_xdt.ExprXDTNamespace.format_localized
-    polars_xdt.ExprXDTNamespace.from_local_datetime
-    polars_xdt.ExprXDTNamespace.is_workday
-    polars_xdt.ExprXDTNamespace.month_name
-    polars_xdt.ExprXDTNamespace.offset_by
-    polars_xdt.ExprXDTNamespace.to_local_datetime
-    polars_xdt.ExprXDTNamespace.to_julian_date
+    polars_xdt.ceil
+    polars_xdt.day_name
+    polars_xdt.format_localized
+    polars_xdt.from_local_datetime
+    polars_xdt.is_workday
+    polars_xdt.month_name
+    polars_xdt.offset_by
+    polars_xdt.to_local_datetime
+    polars_xdt.to_julian_date
