+++ tests/julian_date_test.py
@@ -9,18 +9,19 @@ import pandas as pd
 import polars_xdt as xdt
 
 
-
-
 @given(
-    date=st.datetimes(min_value=dt.datetime(1, 1, 1), max_value=dt.datetime(9999, 12, 31)),
+    date=st.datetimes(
+        min_value=dt.datetime(1, 1, 1), max_value=dt.datetime(9999, 12, 31)
+    ),
 )
 def test_against_pandas(
     date: dt.date,
 ) -> None:
-    df = pl.DataFrame({'a': [date]}, schema={'a': pl.Datetime('ms')})
-    result = df.select(xdt.to_julian_date('a'))['a'].item()
-    expected = pd.Timestamp(df['a'].item()).to_julian_date()
+    df = pl.DataFrame({"a": [date]}, schema={"a": pl.Datetime("ms")})
+    result = df.select(xdt.to_julian_date("a"))["a"].item()
+    expected = pd.Timestamp(df["a"].item()).to_julian_date()
     assert result == expected
+
 
 @given(
     date=st.dates(min_value=dt.date(1, 1, 1), max_value=dt.date(9999, 12, 31)),
@@ -28,7 +29,7 @@ def test_against_pandas(
 def test_against_pandas_date(
     date: dt.date,
 ) -> None:
-    df = pl.DataFrame({'a': [date]})
-    result = df.select(xdt.to_julian_date('a'))['a'].item()
-    expected = pd.Timestamp(df['a'].item()).to_julian_date()
+    df = pl.DataFrame({"a": [date]})
+    result = df.select(xdt.to_julian_date("a"))["a"].item()
+    expected = pd.Timestamp(df["a"].item()).to_julian_date()
     assert result == expected
