+++ tests/test_business_offsets.py
@@ -22,24 +22,28 @@ def get_result(
     date: dt.date,
     dtype: PolarsDataType,
     by: str | pl.Series,
-    **kwargs: Mapping[str, Any],
+    **kwargs: Any,
 ) -> dt.date:
     if dtype == pl.Date:
         result = (
             pl.DataFrame({"ts": [date]})
-            .select(xdt.offset_by('ts', by=by, **kwargs))["ts"]  # type: ignore[arg-type]
+            .select(xdt.offset_by("ts", by=by, **kwargs))["ts"]  # type: ignore[arg-type]
             .item()
         )
     else:
         try:
             result = (
-                pl.DataFrame({"ts": [dt.datetime(date.year, date.month, date.day)]})
+                pl.DataFrame(
+                    {"ts": [dt.datetime(date.year, date.month, date.day)]}
+                )
                 .select(
-                    xdt.offset_by(pl.col("ts")
-                    .dt.cast_time_unit(dtype.time_unit)  # type: ignore[union-attr]
-                    .dt.replace_time_zone(dtype.time_zone)  # type: ignore[union-attr]
-                    , by=by, **kwargs)
-                    .dt.date()
+                    xdt.offset_by(
+                        pl.col("ts")
+                        .dt.cast_time_unit(dtype.time_unit)  # type: ignore[union-attr]
+                        .dt.replace_time_zone(dtype.time_zone),  # type: ignore[union-attr]
+                        by=by,
+                        **kwargs,  # type: ignore[arg-type]
+                    ).dt.date()
                 )["ts"]
                 .item()
             )
@@ -50,7 +54,9 @@ def get_result(
 
 
 @given(
-    date=st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)),
+    date=st.dates(
+        min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)
+    ),
     n=st.integers(min_value=-30, max_value=30),
     weekend=st.lists(
         st.sampled_from(["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]),
@@ -59,7 +65,9 @@ def get_result(
         unique=True,
     ),
     holidays=st.lists(
-        st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)),
+        st.dates(
+            min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)
+        ),
         min_size=1,
         max_size=300,
     ),
@@ -85,7 +93,12 @@ def test_against_np_busday_offset(
     assume(date not in holidays)
     roll = "raise"
     result = get_result(
-        date, dtype, by=function(f"{n}bd"), weekend=weekend, holidays=holidays, roll=roll  # type: ignore[arg-type]
+        date,
+        dtype,
+        by=function(f"{n}bd"),
+        weekend=weekend,
+        holidays=holidays,
+        roll=roll,
     )
     weekmask = [0 if reverse_mapping[i] in weekend else 1 for i in range(1, 8)]
     expected = np.busday_offset(date, n, weekmask=weekmask, holidays=holidays)
@@ -93,7 +106,9 @@ def test_against_np_busday_offset(
 
 
 @given(
-    date=st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)),
+    date=st.dates(
+        min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)
+    ),
     n=st.integers(min_value=-30, max_value=30),
     weekend=st.lists(
         st.sampled_from(["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]),
@@ -102,7 +117,9 @@ def test_against_np_busday_offset(
         unique=True,
     ),
     holidays=st.lists(
-        st.dates(min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)),
+        st.dates(
+            min_value=dt.date(1969, 1, 1), max_value=dt.date(1971, 12, 31)
+        ),
         min_size=1,
         max_size=300,
     ),
@@ -127,7 +144,12 @@ def test_against_np_busday_offset_with_roll(
     roll: Literal["forward", "backward"],
 ) -> None:
     result = get_result(
-        date, dtype, by=function(f"{n}bd"), weekend=weekend, holidays=holidays, roll=roll  # type: ignore[arg-type]
+        date,
+        dtype,
+        by=function(f"{n}bd"),
+        weekend=weekend,
+        holidays=holidays,
+        roll=roll,  # type: ignore[arg-type]
     )
     weekmask = [0 if reverse_mapping[i] in weekend else 1 for i in range(1, 8)]
     expected = np.busday_offset(
@@ -156,7 +178,7 @@ def test_extra_args(by: str, expected: dt.datetime) -> None:
     df = pl.DataFrame({"dates": [start]})
     result = (
         df.with_columns(
-            dates_shifted=xdt.offset_by('dates', by=by)
+            dates_shifted=xdt.offset_by("dates", by=by)
         ).with_columns(end_wday=pl.col("dates_shifted").dt.strftime("%a"))
     )["dates_shifted"].item()
     assert result == expected
@@ -167,7 +189,7 @@ def test_extra_args_w_series() -> None:
     df = pl.DataFrame({"dates": [start] * 2, "by": ["1bd2h", "-1bd1h"]})
     result = (
         df.with_columns(
-            dates_shifted=xdt.offset_by('dates', by=pl.col("by"))
+            dates_shifted=xdt.offset_by("dates", by=pl.col("by"))
         ).with_columns(end_wday=pl.col("dates_shifted").dt.strftime("%a"))
     )["dates_shifted"]
     assert result[0] == dt.datetime(2000, 1, 4, 2)
@@ -181,7 +203,8 @@ def test_starting_on_non_business() -> None:
     df = pl.DataFrame({"dates": [start]})
     with pytest.raises(pl.ComputeError):
         df.with_columns(
-            dates_shifted=xdt.offset_by('dates', 
+            dates_shifted=xdt.offset_by(
+                "dates",
                 by=f"{n}bd",
                 weekend=weekend,
             )
@@ -192,7 +215,8 @@ def test_starting_on_non_business() -> None:
     holidays = [start]
     with pytest.raises(pl.ComputeError):
         df.with_columns(
-            dates_shifted=xdt.offset_by('dates',
+            dates_shifted=xdt.offset_by(
+                "dates",
                 by=f"{n}bd",
                 holidays=holidays,
                 weekend=weekend,
@@ -201,7 +225,10 @@ def test_starting_on_non_business() -> None:
 
 
 def test_within_group_by() -> None:
-    data = {"a": [1, 2], "date": [dt.datetime(2022, 2, 1), dt.datetime(2023, 2, 1)]}
+    data = {
+        "a": [1, 2],
+        "date": [dt.datetime(2022, 2, 1), dt.datetime(2023, 2, 1)],
+    }
     df = pl.DataFrame(data)
 
     result = (
@@ -222,7 +249,11 @@ def test_within_group_by() -> None:
 
 def test_invalid_roll_strategy() -> None:
     df = pl.DataFrame(
-        {"date": pl.date_range(dt.date(2023, 12, 1), dt.date(2023, 12, 5), eager=True)}
+        {
+            "date": pl.date_range(
+                dt.date(2023, 12, 1), dt.date(2023, 12, 5), eager=True
+            )
+        }
     )
     with pytest.raises(pl.ComputeError):
-        df.with_columns(xdt.offset_by('date', "1bd", roll="cabbage"))  # type: ignore[arg-type]
+        df.with_columns(xdt.offset_by("date", "1bd", roll="cabbage"))  # type: ignore[arg-type]
